Move getFieldName and getFieldError to utils
