+++ admin/src/utils/get-field-name.js
@@ -1,0 +1,9 @@
+const getFieldName = name => {
+  if ( name.indexOf( '.' ) !== -1 ) {
+    return name.split( '.' ).slice( 1 ).join( '' );
+  }
+
+  return name;
+};
+
+export default getFieldName;
